backend/backup/pyproject.toml
[tool.poetry]
name = "backup"
version = "0.1.0"
description = "Infrastructure backup service for CRM platform"
authors = ["CRM Platform Team <dev@example.com>"]
readme = "README.md"
packages = [{ include = "backup" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.1" }
apscheduler = "^3.10.4"
boto3 = "^1.34.0"
aio-pika = "^9.4.1"
psycopg = { version = "^3.1.18", extras = ["binary", "pool"] }
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
moto = { version = "^4.2.13", extras = ["s3"] }
testcontainers = { version = "^4.4.1", extras = ["postgresql", "rabbitmq"] }
freezegun = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
